refactor: unify the 6 hand-copied agent text-turn blocks
Extract runTextTurn (server, app/text-turn.ts) — subscribe/capture/send/
waitForIdle/interrupt-on-timeout — reused by inline-ai (generate + classify),
ghost-text, and delegation. 3-variant result (ok / timeout / error) keeps
timeout-vs-throw distinguishable so each caller's own supersession, staleness
strings, and turn-count semantics stay put. Renderer side: extract runManager
(token/supersession lifecycle) + ai-prompts (prompt data), so ai-session.ts
drops 566->419 lines to just the state machine. Behavior preserved 1:1
(delegation + canned-action tests green).
